users can only delete their own blogs

+++ controllers/blogs.js
@@ -6,7 +6,7 @@ const jwt = require('jsonwebtoken')
 blogsRouter.get('/', async (request, response) => {
   const blogs = await Blog.find({}).populate('user', {
     username: 1,
-    name: 1
+    name: 1,
   })
   response.json(blogs)
 })
@@ -27,13 +27,18 @@ blogsRouter.post('/', async (request, response) => {
 
   const blog = new Blog({
     user: user.id,
-    ...request.body
+    ...request.body,
   })
   const savedBlog = await blog.save()
   response.status(201).json(savedBlog)
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
+  const user = jwt.verify(request.token, process.env.SECRET)
+  const blog = await Blog.findById(request.params.id)
+  if (!blog) return response.status(204).end()
+  if (blog.user.toString() !== user.id.toString())
+    return response.status(401).json({ error: 'token invalid' })
   await Blog.findByIdAndDelete(request.params.id)
   response.status(204).end()
 })
